Separate montage methods from vendor-specific code
The `montage_grid_image_YX` and `montage_stage_pos_image_YX` functions
can be used independently from any input format,
so let's keep them in a separate module.
